+++ sydtest/src/Test/Syd/Path.hs
@@ -21,20 +21,19 @@ tempDirSetupFunc ::
   -- | Temporary directory name template
   String ->
   SetupFunc (Path Abs Dir)
-tempDirSetupFunc template = makeSimpleSetupFunc $ withSystemTempDir template
+tempDirSetupFunc template = SetupFunc $ withSystemTempDir template
 
 tempBinaryFileWithContentsSetupFunc ::
   -- | Temporary directory name template
   String ->
   ByteString ->
   SetupFunc (Path Abs File)
-tempBinaryFileWithContentsSetupFunc template contents =
-  makeSimpleSetupFunc $ \func ->
-    withSystemTempFile
-      template
-      ( \af h -> do
-          SB.hPut h contents
-          hFlush h
-          hClose h
-          func af
-      )
+tempBinaryFileWithContentsSetupFunc template contents = SetupFunc $ \func ->
+  withSystemTempFile
+    template
+    ( \af h -> do
+        SB.hPut h contents
+        hFlush h
+        hClose h
+        func af
+    )